operator: Fix deletion of newer NMstate CRs (#816)
Operator deletes new NMState CR if there is already one
existing. The problem is however, that current CRs are not
always listed in correct order, which means that the current
logic sometimes doesn't delete new CR when another already
exists, causing e2e tests to flake.

This commit sorts listed NMstate CRs when there are more than one
of them, and compare the name of the reconciled CR with the oldest.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>