e2e: Remove rollback test that is now obsolete (#832)
In previous commit that removed vlan-filtering script,
this test was changed to use bad yaml, instead of renaming
the vlan-filtering binary, to make applyDesiredState fail
on purpose.

The problem is, this change is not equivalent - previously, the
desired state was applied, and only then the function failed.
Now however, this test fails to apply the desired state, so rollback
is not really exercised.

The thing is, after removing the use of the vlan-filtering script,
the only function call that can potentially fail is nmstatectl commit,
and there doesn't seem to be practical way to make it fail.

For that reason, this test appears to be obsolete and can be removed.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>