e2e: wait for re-reconcile after node reboot (#844)
Instead of waiting for update of LastSuccessfulUpdate of NNS,
wait for enactment to become progressing and then policy to
be Available again.

The reason we can't just wait for the timestamp update after
node reboot is that when a node boots up, the policy
reconcile may happen before NNS update, which means that the
NNS still contains the state with the policy configuration.
In this case, there is no change in the node state, which
means NNS is not updated (because there's no difference).

With that said, we still need to make sure we wait for the policy
is reconciliateion after reboot, before verifying the configuration
is present. This change uses the enactment and policy condition
transitions instead of the timestamp.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>