Remove leader election (#912)
* cert-manager: Remove leader election

Now that the cert-manager is living at it's own pod and we have just
one replica with Recreate upgrade strategy we don't need leader election
since is going to be only one active.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* webhook: Remove leader election

Now that cert-manager is not living inside the webhook and only one
webhook receive the request we don't need leader election.

Signed-off-by: Quique Llorente <ellorent@redhat.com>