Don't wait for node boot when waiting for nncp transition
The function waiting for node reboot is checking if ssh connects
every 5 seconds, which seems to be enough for a simple policy to
start reconciling and even finish, meaning that subsequent
check, that waits for a NNCP transition timestamp to be updated
fails (as no more transitions happen).

This change doesn't wait for the node to boot and relies only on
the transition timestamp update.

Also, fixed restartNodeWithoutWaiting which didn't return an error

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>