Fix issues in DNS tests (#929)
Addresses https://github.com/nmstate/kubernetes-nmstate/issues/927

Following issues are causing the tests to fail:
1. removing of IP address from an interface was done by using
`absent` withing ipv4 subtree. This is likely a bug in nmstate
as this subtree doesn't specify such property.
2. in static_addr_and_route_test, one IP address was assigned
to all worker nodes. This change designates only one worker
node to have the NNCPs configured. This reduction
doesn't limit our coverage as it shouldn't matter if this configuration
is applied to one node or more.
3. some policies missed ipv4/6.enabled: true, which may not be causing
an issue, but this change specifies the enabled property explicitly.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>