Add hack utility for sorting handler logs (#932)
It's not uncommon to get a log file that contains
concatenated logs of cluster handler pods, meaning
the log events are not ordered by time.
It's  impossible to process and investigate issues in such logs
before ordering the events by time.

Adding a python script that orders such a file.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>