webhook: Deny NNCP update if capture is modified
Updateing the capture field at NNCP can break scenarios where nodes are
rebooted and NetworkManager configuration is not presistent. This change
prevent that but just disallow update capture field, if it need to be
modified the NNCP has to be re-created.

This is a first step before implementing a better solution.

Signed-off-by: Quique Llorente <ellorent@redhat.com>