build: Use multistage to build the manager
In the future the project is going to be integrated with the rust
version from nmstate, this integration is going to be done using CGO and
for that we need to install the nmstate rust libraries and headers. This
change transition to multistage so the nmstate-devel library can be
installed at the build stage.

Signed-off-by: Quique Llorente <ellorent@redhat.com>