[release-0.64] backport recent fixes (#974)
* deps: Pin NetworkManager to 1.34 (#958)

We have found some issues with NetworkManager 1.36 that we are
debugging, while issues are fixed this change pin NetworkManager to
1.34.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Fix periodic re-reconciles of NNCPs with name equal to node (#957)

Policies that are named the same as worker nodes are getting
periodically re-reconciled.
It seems that the For(&corev1.Node{}) creates a watch of it's own
that overrides the builder.WithPredicates(onLabelsUpdatedForThisNode)
defined in the subsequent Watches(...) statement.

This commit creates the controller manually without using
controller-runtime builder.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>

* e2e: don't wait for reboot when testing default bridge IP (#959)

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>

* Fix ignoring the error if the probe fails during rollback (#960)

If the probe also fails during the rollback, a nil error is
returned and hence the operation/nnce is marked as successful
although  the probe failed while applying the conf and during
the rollback.

Signed-off-by: Nijin Ashok <nashok@redhat.com>

* operator: Add "resources.requests" for CPU and Mem (#971)

Following best practices for pods configuration and the status quo at
handler. This change ad the resources requests for CPU and memory
following the output of `crictl stats` on the operator container.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* operator: Use priority class cluster critical (#972)

The operator can be replaced by a pod running at different node. This
change represent that by setting priority class at operator to
`system-cluster-critical`.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

Co-authored-by: Enrique Llorente Pastora <ellorent@redhat.com>
Co-authored-by: Nijin Ashok <nashok@redhat.com>