Don't involve notReady nodes in status calculation (#985)
In big clusters, it's possible that some nodes are not ready, either
due to maintenance or networking issue. This change excludes NotReady
nodes from NNCP Status calculation, so that policy is considered
as Available, when it successfully applies to all Ready nodes.

When NotReady node turns Ready, it re-reconciles and updates the policy
accordingly.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>