Merge NNCP controller definition (#1006)
As we declare two NNCP controllers (one for changes on NNCP and one for
changes on nodes (node labels)), there is a change that Reconcile runs in
parallel. This can cause issues if the same policy is getting reconciled
in parallel on the same node.
This commit addresses it and merges both definitions of the NNCP
controllers so we have only one Reconcile run at a time (by default).

Signed-off-by: Christoph Stäbler <cstabler@redhat.com>