build: Use go cross compiling instead of qemu (#1009)
* build: Use go cross compiling instead of qemu

The buildx build command start a QEMU instance if the container arch is
not the one on the host, go toolchain has cross compiling functionality
so qemu is not needed for the compiling stuff. This change force the
compile stage to be amd64 since it's the arch at CI and then runs cross
compiling to compile amd64 or arm64 so it does not compile over qemu.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* build: Pin to go.mod go version

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* build: Reuse Dockerfile for future nmstate pin

Signed-off-by: Quique Llorente <ellorent@redhat.com>