Set policy status in reconcile if it's Unknown (#1041)
This change addresses a reported BZ that Policy has Unknown
status when it's already progressing. That is because Reconcile
loop updates policy status only when it's finished.

This doesn't look good, especially on single node cluster, where
policy is never progressing, despite the enactment is progressing.

To limit API hits, the update is only attempted if the policy has
Available condition set to Unknown.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>