e2e,examples: fix e2e examples
Currently, examples test only the very last example yaml, which is
'worker selector'. The reason for Ginkgo's multistep flow, which
1. collects all the container nodes (Context, Describe, When) and stores
the runtime blocks (the actual test functions passed to It specs) for
later execution.
2. executes those blocks

This means that in our case, the containers have correct names,
generated during the loop, but the functions all hold reference
to the last example, so all specs work with worker-selector.yaml.

Fixing this by storing to a local variable.

This change also includes fixes to the examples, that are not working
properly. The DNS test in particular requires proper cleanup.
This change adds cleanupState field to the exampleSpec to perform custom
cleanup at AfterEach stage.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>