Respect new node label (node-role.kubernetes.io/control-plane) on control plane nodes (#1103)
* Use new node selector for cp nodes (node-role.kubernetes.io/control-plane)

Signed-off-by: Christoph Stäbler <cstabler@redhat.com>

* Add node affinity to respect both control-plan/master node labels

Adds default node affinity for infra pods to schedule them on
control-plane nodes having the new
(node-role.kubernetes.io/control-plane) label or on the nodes with the
old (node-role.kubernetes.io/master) label, as OR can't be used in a
nodeSelector.
This commit can be reverted, as soon as the control-plane nodes do not
have the old (node-role.kubernetes.io/master) label anymore.

Signed-off-by: Christoph Stäbler <cstabler@redhat.com>