tmp: Skip ovs-bridge e2e test with nmstate 1.2
An OVS-bridge e2e tests was reenabled in
a5ed4b58c8efc3aec24f2bf4637d85f848fc22ea in upstreams. This required a
fix in nmstate which was not yet backported to nmstate 1.2. Therefor we
keep this test skipping until it was backported to nmstate 1.2 as this
does not make things worse.

Signed-off-by: Christoph Stäbler <cstabler@redhat.com>