e2e,upgrade: fall back to current examples if needed (#1120)
When testing upgrade, example yamls from previous tag
are downloaded. This causes an issue when adding a new
example yaml, because it doesn't yet exist in previous tag.

This change falls back to current doc/examples path if
an example yaml is not found in previous tag.
This allows to smoothly add new example yamls.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>