Probes: choose the default gw from the main routing table (#1153)
* Probes: choose the default gw from the main routing table

The current logic relies on the first matching route entry with
destination == 0.0.0.0/0, in order to find the default gateway to ping.
When multiple routing table exist, we may have multiple entries with
that destination, some of them may not be usable and even if they were,
pinging that particular gateway doesn't assure the health of the node.

To overcome this, we narrow the filter to pick the next hop of the
default route for the "main" routing table.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

* E2E: add a test where we change the default gw for different table

We want to ensure ping still works if we change the default gateway for
a routing table different from main one.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

---------

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>