handler: Remove nmcli vlan-filtering activation (#1171)
For upgrade purposes the vlan filtering was being configured with nmcli
before apply the modified state so the old ip link version of k-nmstate
will not fail on nmstate verification. This change remove the nmcli
command to configure vlan filterind and also the nmcli versioning at NNS
since we no longer need it, it also remove NetworkManager from the
container.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>