Enable ipv4 and ipv6 on primary nic when resetting
By default NMState disables ipv4 and ipv6 if they are not explicitly
set to enabled. This can cause the reset policy to break network
connectivity on nodes since the primary interface may lose all of
its addresses. To avoid that, set both ipv4 and ipv6 to enabled with
DHCP so we will always configure them with an address.

Signed-off-by: Ben Nemec <bnemec@redhat.com>