handler: Propertly filter unmanaged veth (#1181)
The number of veth interfaces not managed by NetworkManager at a k8s
cluster is massive since they are all the network interfaces of
containers running at the node and it pollute the NodeNetworkState
without much use, to fix that there is a code that ignore those. At some
envs the "veth" parth is not present since the peer is at different
namespace so the detection logic was failing. This change goes back to
use "type == veth" now that the veth type is no longer "ethernet".

Signed-off-by: Enrique Llorente <ellorent@redhat.com>