UPSTREAM: <carry>: Drop machine-config workaround from e2e tests
Thanks to some changes upstream in the e2e test suite we no longer
need to do this workaround. Waiting for the entire cluster to reboot
is now a waste of time and likely contributing to the timeouts that
occur on a regular basis.

(cherry picked from commit 53dab2ecd6cb19ce533610c54fb1b9dd87804c62)
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>