test: pass WithStatusSubresource to the fake client
With the upgrade to controller-runtime v0.15 we need to supplement the
fake builder to contain instantiation `WithStatusSubresource` in order
to be able to update status of the object that we have created.

In previous versions of controller-runtime this was not necessary
because the semantics was ignored but in v0.15 the fake client cannot
infer the status subresource without its explicit instantiation.

Ref.: https://github.com/kubernetes-sigs/controller-runtime/issues/2362

Signed-off-by: Mat Kowalski <mko@redhat.com>