e2e: Use metrics container when executing commands
This PR adds explicit `-c nmstate-metrics` to the command invocations
that use `kubectl.sh exec [...]`. This should not change the current
behaviour as by default we execture them in this container, however it's
been noticed that the following stderr is produced

```
stderr: Defaulted container "nmstate-metrics" out of: nmstate-metrics, kube-rbac-proxy
```

so for being explict, we are defining where to execute the command.

Signed-off-by: Mat Kowalski <mko@redhat.com>