e2e: don't require empty starting state for bridge mappings (#1248)
This commit changes the behaviour of the OVN bridge mapping test so that
we do not require the initial state to be empty. Thanks to this we make
e2e tests working correctly on machines whre OVN is already used and
bridges are deployed.

Without this change the test will pass only if no OVN bridge mappings
exist on the node. The assumption is quite strong and not really
desired.

Signed-off-by: Mat Kowalski <mko@redhat.com>