test: ability to use kubectl instead of ssh
Right now the test framework uses SSH to access nodes in order to run
e.g. nmstatectl commands. With this change we add ability to use kubectl
debug functionality to run commands directly on the node.

This is useful when running test suite in clusters with no direct SSH
access to the nodes, e.g. remote cloud clusters and not a local bare
metal setups.

In order to use this feature, make sure the `USE_KUBECTL_DEBUG_NODE` env
variable is set when calling Makefile. Also make sure that `KUBECTL`
points to the kubectl binary that will be used when making the calls.

Signed-off-by: Mat Kowalski <mko@redhat.com>