test: add DNS tests without interface configuration
This PR adds a set of tests where we configure DNS servers without
providing any per-interface configuration. This is a canonical way to
use k-nmstate to provide global DNS configuration. It is useful in
environments where we are not allowed to modify network interfaces, but
we still want to reconfigure DNS.

Signed-off-by: Mat Kowalski <mko@redhat.com>