OCPBUGS-44596: Mount OVS directory instead of socket file
When mounting the OVS DB socket directly as a file
(`/run/openvswitch/db.sock`) instead of mounting the whole directory,
handler container loses access to the OVS DB in case of OVS daemon
restart on the host. This can only be fixed by restarting the container
so that it mounts again the correct file.

To remediate this, we are mounting the whole `/run/openvswitch`
directory so that DB connection can be gracefully restored by the
handler itself.

Closes: OCPBUGS-44596
Signed-off-by: Mat Kowalski <mko@redhat.com>