OPNET-614: Configurable DNS probe (#1297)
* Extend logs collected after CI

Signed-off-by: Mat Kowalski <mko@redhat.com>

* OPNET-632: Extend nmstate CRD to include DNS probe configuration

This PR extends the nmstate CRD so that the custom DNS host can be
configured using `.Spec.ProbeConfiguration.DNS.Host` path.

In the future this configuration will be used by the handler pod when
running its healthchecks.

Signed-off-by: Mat Kowalski <mko@redhat.com>

* OPNET-633: Implement custom DNS probe handler

This commmit implements handling of the PROBE_DNS_HOST variable
configured via the nmstate CR.

By default it keeps the old behaviour of checking NS record for
root-servers.net domain.

In case custom domain name is configured, it will use A or AAAA record.

Signed-off-by: Mat Kowalski <mko@redhat.com>

---------

Signed-off-by: Mat Kowalski <mko@redhat.com>