nmstatectl: context for command and timeout for Run
This commit adds a context to the function executing nmstatectl on the
host. Thanks to this, we can add timeout of 10 seconds to every
execution of `nmstatectl show`.

Given this is a reasonable time in which this should finish, we will
prevent process from hanging in rare cases when `show` takes infinite
amount of time to finish.

Signed-off-by: Mat Kowalski <mko@redhat.com>