OPNET-673: Implement default network policy (#1337)
This PR implements a default set of network policies that the operator
will deploy. This is supposed to increase security of the cluster so
that rogue workload in the operator namespace has limited access to the
outside world.

The PR adds a set of policies for the operator container, webhook,
metrics and console plugin. It adds a default "deny-all" policy for the
whole namespace. As handler pod runs as privileged workload with
host-networking, it is not necessary to implement policies there as they
would not be applied anyway.

Signed-off-by: Mat Kowalski <mko@redhat.com>