netpol: use targeted policy for deny-all
This commit changes the deny-all policy from targetting the whole
namespace to targetting only labelled workload. This will help with
cases when operator runs in a shared namespace and not its own unique.

We require this because with OLM deployment method we do not really
forbid users from using a shared namespace.

Signed-off-by: Mat Kowalski <mko@redhat.com>