operator: use strategic merge to prevent losing labels (#1357)
We are changing MergeFrom to StrategicMergeFrom so that when applying
manifests we do not lose any content.

When using MergeFrom, if the desired namespace contains a custom label
or custom annotation it will be wiped because we set namespace labels.
In this case the lists are not merged but forcefully overriden.

When using StrategicMergeFrom, lists are correctly merged. As a result,
custom annotations are not removed.

In principle this apply to every field that is a list, not only to the
namespace labels and annotations.

Signed-off-by: Mat Kowalski <mko@redhat.com>