crd: implement NMState CRD status subresource (#1334)
* cluster: Add allowerasing when upgrading NM

Signed-off-by: Enrique Llorente <ellorent@redhat.com>

* cluster: Wait for nmstate status

Signed-off-by: Enrique Llorente <ellorent@redhat.com>

* operator: implement NMState CRD status subresource

This commit implements comprehensive status reporting for the NMState operator
with proper condition management and resource reconciliation.

Asisted-by: claude code, gemini
Signed-off-by: Enrique Llorente <ellorent@redhat.com>

* operator: Use server side apply

Signed-off-by: Enrique Llorente <ellorent@redhat.com>

* operator: Handle missing CRDs gracefully

Skip resources when their CRDs are not available instead of failing,
logging a warning message for better debugging.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>

* operator: Pass Reconcile ctx down

As part of the operator refactoring to be idempotent it was detected
that some of the underlaying functions use a new context.TODO,
context.Background, those should use the ctx from Reconcile. This change
address that by passing it down

Signed-off-by: Enrique Llorente <ellorent@redhat.com>

---------

Signed-off-by: Enrique Llorente <ellorent@redhat.com>