OCPBUGS-57067: Remove arch node selector
This product supports more than a single architecture. Yet, the code is
written in a way that it will only deploy pods on the architecture that
matches the architecture of the first node where the controller
container is running.

As a result, it does not work in multi-arch clusters. If your controller
got scheduled in amd64 node, all the containers will only run on amd64
nodes. Similarly if you got scheduled first in arm, it only runs on arm.

This change removes the architecture node selector completely. As a
result, we will try scheduling on every node. There may be a side effect
that if you have cluster with some exotic nodes they will not run
successfully. But if this is the case, NMstate CR has ability to write a
custom node selector.

Signed-off-by: Mat Kowalski <mko@redhat.com>