Prevent non mandatory mutating webhooks from making network configuration fail. (#1387)
* fix(controller): Prevent NNCP reconciler to trigger status webhook

Reseting the NNCP status by cleaning up the status condition list force
the NNCP status webhook to be called and set conditions to Unknown. This
change directly set the NNCP status conditions to unknown at the
controller.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>

* fix(webhook): Ignore errors on non validating ones

The webhook logic that change reset the NNCP to state that it's changing
should not stop the network state from being applied if there are some
errors like TLS not ready, to do that this change set the errorPolicy:
Ignore on those ones

Signed-off-by: Enrique Llorente <ellorent@redhat.com>

---------

Signed-off-by: Enrique Llorente <ellorent@redhat.com>