License cleanup
This commit aligns LICENSE file at the root of the repo with licenses of
all the other files.

The one at the root of the repo states (apparently wrongly) GPL-2.0 but
all the files containing actual code use Apache-2.0 header since very
beginning.

To avoid the confusion and make it clear that this is an Apache-2.0
project, we are fixing the LICENSE file at the root of the repo.

Mandatory INAL but I am doing this change as a technical cleanup, not
as a legal action to relicense. Source code of this project already
lives in the files with Apache-2.0 headers.

Signed-off-by: Mat Kowalski <mko@redhat.com>