metrics: introduce metrics based on the NNS
This PR introduces a small set of metrics based on the NodeNetworkState,
i.e. `kubernetes_nmstate_network_interfaces` and
`kubernetes_nmstate_routes`. Those track number of network interfaces
and number of routes for every node in the cluster where nmstate handler
pods are running.

This creates a skeleton for k-nmstate to create metrics based on the
effective state of the network on the system and not only based on the
content of NNCP.

In the future it may be used to create statistics about network
topologies of clusters.

Signed-off-by: Mat Kowalski <mko@redhat.com>