handler: fetch fresh NNCP before creating NNCE
This PR is an attempt to remediate a race with `Enactment
node02.ovs-wrong-ip not deleted`. In a hypothetical scenario of outdated
policy cache in one of the handler nodes it could decide to create an
enactment for the node even though a fresh policy is not targetting this
node (by having a narrow node selector).

Signed-off-by: Mat Kowalski <mko@redhat.com>