test: add cleanup in BeforeEach to avoid duplicated policies
This change attempts to remediate an issue where test fails because of a
duplicated policy, i.e.

```
nodenetworkconfigurationpolicies.nmstate.io "test-policy" already exists
```

which may potentially be an outcome of a missing cleanup in some other
test. With this change we will always try to create the policy first
before starting a test to ensure a clean starting state.

Signed-off-by: Mat Kowalski <mko@redhat.com>